generates the new sdk patched
