+++ plenigo-client/plenigo/models/__init__.py
@@ -3,7 +3,7 @@
 from .access_right_data import AccessRightData
 from .access_right_item_data import AccessRightItemData
 from .access_right_item_data_access_right_data import AccessRightItemDataAccessRightData
-from .access_right_item_data_data_item import AccessRightItemDataDataItem
+from .access_right_item_data_data import AccessRightItemDataData
 from .access_rights_data import AccessRightsData
 from .additional_chain_data import AdditionalChainData
 from .additional_chain_data_data import AdditionalChainDataData
@@ -237,7 +237,7 @@ __all__ = (
     "AccessRightData",
     "AccessRightItemData",
     "AccessRightItemDataAccessRightData",
-    "AccessRightItemDataDataItem",
+    "AccessRightItemDataData",
     "AccessRightsData",
     "AdditionalChainData",
     "AdditionalChainDataData",
